Changet parameter type in GetById method.
This is done due to the fact that the IDs in the database are represented by the long type.

+++ TrainSchedule/RepositoryInterfaces/IRepository.cs
@@ -17,6 +17,6 @@ namespace TrainSchedule.RepositoryInterfaces
         /// </summary>
         /// <param name="id">ID of object to get</param>
         /// <returns>Recieved object</returns>
-        T GetById(uint id);
+        T GetById(long id);
     }
 }
